extract profit distribution
