+++ src/services/sale/set-sale-status.ts
@@ -6,6 +6,7 @@ import { OrganizationRepository } from '@/repositories/organization-repository'
 import { ProfilesRepository } from '@/repositories/profiles-repository'
 import { UnitRepository } from '@/repositories/unit-repository'
 import { PaymentStatus, TransactionType } from '@prisma/client'
+import { distributeProfits } from './profit-distribution'
 
 interface SetSaleStatusRequest {
   saleId: string
@@ -63,62 +64,11 @@ export class SetSaleStatusService {
           transaction: { connect: { id: transaction.id } },
         })
 
-        const org = await this.organizationRepository.findById(
-          user?.organizationId as string,
-        )
-        if (!org) throw new Error('Org not found')
-        const barberTotals: Record<string, number> = {}
-        let ownerShare = 0
-        for (const item of updatedSale.items) {
-          const value = item.price ?? 0
-          if (item.product) {
-            ownerShare += value
-          } else if (item.barberId) {
-            const perc = item.porcentagemBarbeiro ?? 100
-            const valueBarber = (value * perc) / 100
-            barberTotals[item.barberId] =
-              (barberTotals[item.barberId] || 0) + valueBarber
-            ownerShare += value - valueBarber
-          } else {
-            ownerShare += value
-          }
-        }
-        for (const [barberId, amount] of Object.entries(barberTotals)) {
-          const userBarber = updatedSale.items.find(
-            (item) => item.barber?.id === barberId,
-          )
-          if (!userBarber) throw new Error('Barber not found')
-          if (
-            userBarber &&
-            userBarber.barber &&
-            userBarber.barber.profile &&
-            userBarber.barber.profile.totalBalance < 0
-          ) {
-            const balanceBarber = userBarber.barber.profile.totalBalance
-            const valueCalculated = balanceBarber + amount
-            if (valueCalculated <= 0) {
-              await this.unitRepository.incrementBalance(
-                updatedSale.unitId,
-                amount,
-              )
-            } else {
-              await this.unitRepository.incrementBalance(
-                updatedSale.unitId,
-                balanceBarber * -1,
-              )
-              await this.organizationRepository.incrementBalance(
-                org.id,
-                balanceBarber * -1,
-              )
-            }
-          }
-          await this.profileRepository.incrementBalance(barberId, amount)
-        }
-        await this.unitRepository.incrementBalance(
-          updatedSale.unitId,
-          ownerShare,
-        )
-        await this.organizationRepository.incrementBalance(org.id, ownerShare)
+        await distributeProfits(updatedSale, user?.organizationId as string, {
+          organizationRepository: this.organizationRepository,
+          profileRepository: this.profileRepository,
+          unitRepository: this.unitRepository,
+        })
 
         return { sale: updatedSale }
       } catch (error) {
